backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Alembic==1.13.1
Marshmallow==3.20.1
Flask-Admin==1.6.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# Email service
sendgrid==6.11.0

# WSGI server for production
gunicorn==21.2.0

# Production database support
psycopg2-binary==2.9.9

# File handling and utilities
Pillow==10.1.0
Werkzeug==3.0.1

# Development tools (optional in production)
ruff==0.14.2
black==25.9.0